Fix : 자산 배분 설정 분리
